# app-backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Babble API",
    "version": "1.0.0",
    "description": "Babble API — 발음 연습 서비스"
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ApiErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string"
          },
          "errorCode": {
            "type": "string"
          },
          "errorKey": {
            "type": "string"
          }
        }
      },
      "PaginatedResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {}
          },
          "total": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "offset": {
            "type": "integer"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string"
          },
          "errorCode": {
            "type": "string"
          },
          "errorKey": {
            "type": "string"
          }
        }
      },
...
      "CreateGameSessionDto": {
        "type": "object",
        "required": [
          "gameType",
          "difficulty",
          "correctCount",
          "totalCount",
          "duration",
          "score"
        ],
        "properties": {
          "gameType": {
            "type": "string",
            "enum": [
              "WORD_MATCH",
              "PRONUNCIATION_QUIZ",
              "SPEED_READ"
            ]
          },
          "difficulty": {
            "type": "string",
            "enum": [
              "EASY",
              "MEDIUM",
              "HARD"
            ]
          },
          "correctCount": {
            "type": "integer",
            "minimum": 0
          },
          "totalCount": {
            "type": "integer",
            "minimum": 1
          },
          "duration": {
            "type": "integer",
            "minimum": 0,
            "description": "소요 시간 (초)"
          },
          "score": {
            "type": "integer",
            "minimum": 0
          },
          "comboMaxStreak": {
            "type": "integer",
            "minimum": 0,
            "description": "최대 콤보 연속 정답 수"
          },
          "wordResults": {
            "type": "array",
            "description": "단어별 결과 (v2 XP 차등 지급용)",
            "items": {
              "$ref": "#/components/schemas/WordResultDto"
            }
          }
        }
      },
...
      "WordResultDto": {
        "type": "object",
        "required": [
          "scriptId",
          "word",
          "wordIndex",
          "isCorrect",
          "attempts",
          "hintUsed"
        ],
        "properties": {
          "scriptId": {
            "type": "integer",
            "minimum": 1
          },
          "word": {
            "type": "string",
            "minLength": 1
          },
          "wordIndex": {
            "type": "integer",
            "minimum": 0
          },
          "isCorrect": {
            "type": "boolean"
          },
          "attempts": {
            "type": "integer",
            "minimum": 1
          },
          "hintUsed": {
            "type": "boolean"
          }
        }
      },
...
      "CreateScriptDto": {
        "type": "object",
        "required": [
          "title",
          "content"
        ],
        "properties": {
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "difficulty": {
            "type": "string",
            "enum": [
              "EASY",
              "MEDIUM",
              "HARD"
            ]
          },
          "articulationPlace": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "chapterId": {
            "type": "integer",
            "nullable": true
          },
          "orderIndex": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
...
      "Script": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "difficulty": {
            "type": "string",
            "enum": [
              "EASY",
              "MEDIUM",
              "HARD"
            ]
          },
          "articulationPlace": {
            "type": "string",
            "enum": [
              "BILABIAL",
              "ALVEOLAR",
              "VELAR",
              "PALATAL",
              "GLOTTAL",
              "MIXED",
              "NONE"
            ]
          },
          "targetConsonants": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/assessments": {
      "post": {
        "tags": [
          "진단(Assessment)"
        ],
        "summary": "음성 파일 업로드 및 진단 생성",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "audio": {
                    "type": "string",
                    "format": "binary"
                  },
                  "scriptId": {
                    "type": "integer",
                    "example": 1
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "업로드 성공 및 진단 생성 완료",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "파일이 없거나 잘못된 형식",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "인증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "진단(Assessment)"
        ],
        "summary": "진단 이력 조회",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "type": "integer",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse"
                }
              }
            }
          },
          "401": {
            "description": "인증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/game-sessions": {
      "post": {
        "tags": [
          "게임(Game)"
        ],
        "summary": "게임 세션 생성",
        "description": "게임 플레이 결과를 기록합니다.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGameSessionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "생성 성공"
          },
          "401": {
            "description": "인증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "게임(Game)"
        ],
        "summary": "게임 세션 이력 조회",
        "description": "사용자의 게임 플레이 이력을 조회합니다.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "type": "integer",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "조회 성공"
          },
          "401": {
            "description": "인증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/game-configs": {
      "get": {
        "tags": [
          "어드민-게임설정(Admin-GameConfig)"
        ],
        "summary": "전체 게임 설정 목록",
        "description": "모든 게임 설정을 반환합니다. (어드민 전용)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "key": {
                            "type": "string"
                          },
                          "value": {},
                          "description": {
                            "type": "string"
                          },
                          "category": {
                            "type": "string"
                          },
                          "updatedAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "updatedBy": {
                            "type": "integer",
                            "nullable": true
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "인증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "관리자 권한 필요",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/scripts": {
      "get": {
        "tags": [
          "구절(Script)"
        ],
        "summary": "구절 목록 조회",
        "description": "발음 진단용 구절 목록을 조회합니다. 필터링 옵션 지원.",
        "parameters": [
          {
            "in": "query",
            "name": "category",
            "schema": {
              "type": "string"
            },
            "description": "카테고리 (bible, practice, daily 등)"
          },
          {
            "in": "query",
            "name": "difficulty",
            "schema": {
              "type": "string",
              "enum": [
                "EASY",
                "MEDIUM",
                "HARD"
              ]
            },
            "description": "난이도"
          },
          {
            "in": "query",
            "name": "articulationPlace",
            "schema": {
              "type": "string",
              "enum": [
                "BILABIAL",
                "ALVEOLAR",
                "VELAR",
                "PALATAL",
                "GLOTTAL",
                "MIXED",
                "NONE"
              ]
            },
            "description": "조음 위치"
          },
          {
            "in": "query",
            "name": "isActive",
            "schema": {
              "type": "boolean"
            },
            "description": "활성화 상태"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 50
            },
            "description": "조회할 개수"
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "type": "integer",
              "default": 0
            },
            "description": "오프셋"
          }
        ],
        "responses": {
          "200": {
            "description": "조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "items": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Script"
                          }
                        },
                        "total": {
                          "type": "integer"
                        },
                        "limit": {
                          "type": "integer"
                        },
                        "offset": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "구절(Script)"
        ],
        "summary": "구절 생성",
        "description": "새로운 발음 진단 구절을 생성합니다.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateScriptDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "생성 성공"
          },
          "400": {
            "description": "필수 필드 누락"
          },
          "401": {
            "description": "인증 실패"
          }
        }
      }
    },
...
  },
...
}